+++ airesources/C++/hlt/entity_id.cpp
@@ -1,0 +1,67 @@
+#include "entity_id.hpp"
+
+namespace hlt {
+    EntityId::EntityId() {
+        type = EntityType::InvalidEntity;
+        _player_id = -1;
+        _entity_index = -1;
+    }
+
+    auto EntityId::is_valid() const -> bool {
+        return type != EntityType::InvalidEntity &&
+               _player_id >= -1 && _entity_index >= 0;
+    }
+
+    auto EntityId::invalid() -> EntityId {
+        return EntityId();
+    }
+
+    auto EntityId::player_id() const -> PlayerId {
+        return static_cast<PlayerId>(_player_id);
+    }
+
+    auto EntityId::entity_index() const -> EntityIndex {
+        return static_cast<EntityIndex>(_entity_index);
+    }
+
+    auto EntityId::for_planet(EntityIndex index) -> EntityId {
+        auto result = EntityId();
+        result.type = EntityType::PlanetEntity;
+        result._entity_index = static_cast<int>(index);
+        return result;
+    }
+
+    auto EntityId::for_ship(PlayerId player_id, EntityIndex index) -> EntityId {
+        auto result = EntityId();
+        result.type = EntityType::ShipEntity;
+        result._player_id = player_id;
+        result._entity_index = static_cast<int>(index);
+        return result;
+    }
+
+    auto operator<<(std::ostream &ostream,
+                    const EntityId &id) -> std::ostream & {
+        switch (id.type) {
+            case EntityType::InvalidEntity:ostream << "[Invalid ID]";
+                break;
+            case EntityType::PlanetEntity:
+                ostream << "[Planet " << id.entity_index() << "]";
+                break;
+            case EntityType::ShipEntity:
+                ostream << "[Ship " << static_cast<int>(id.player_id());
+                ostream << ' ' << id.entity_index() << "]";
+                break;
+        }
+
+        return ostream;
+    }
+
+    auto operator==(const EntityId &id1, const EntityId &id2) -> bool {
+        return id1._player_id == id2._player_id
+               && id1._entity_index == id2._entity_index;
+    }
+
+    auto operator!=(const EntityId &id1, const EntityId &id2) -> bool {
+        return !(id1 == id2);
+    }
+}
