--- airesources/C++/hlt/move.cpp
@@ -1,5 +1,0 @@
-//
-// Created by David Li on 7/19/17.
-//
-
-#include "move.hpp"
