+++ airesources/C++/hlt/navigation.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+namespace hlt {
+}
