+++ airesources/C++/hlt/types.hpp
@@ -1,7 +1,9 @@
 #pragma once
 
+#include <unordered_map>
+
 namespace hlt {
-    /** Uniquely identifies each player. */
+    /// Uniquely identifies each player.
     typedef int PlayerId;
 
     /**
@@ -12,4 +14,11 @@ namespace hlt {
      * EntityIndex alone.
      */
     typedef unsigned int EntityIndex;
+
+    template<typename T>
+    using entity_map = std::unordered_map<EntityIndex, T>;
+
+    /// A poor man's std::optional.
+    template<typename T>
+    using possibly = std::pair<T, bool>;
 }
